+++ CHANGELOG.md
@@ -8,7 +8,7 @@
     - [Added] Added a new method to retrieve all annotations associated with a data product uuid from metastore.
     - [Added] Added a GET method to get an annotation from metastore by id.
     - [Added] Added a GET method to get all annotations associated with data product uuid from metastore.
-    - [Changed] Updated the DataProductAnnotations class to all for conversion from database results.
+    - [Changed] Updated the DataProductAnnotations class to allow for conversion from database results.
 
 - [NAL-1295](https://jira.skatelescope.org/browse/NAL-1295)
     - [Added] Added a class to represent the new annotations table.